clarify the configuration in for storing traces in automations
It is not clear from the documentation what setting `stored_traces` to `1` means. Does it mean "keep all traces" or "keep one trace"? Assuming it means "keep one trace", putting a value of 10 in the documentation makes it more clear that this integer value represents the number of traces to be stored.